requirements.txt
# Add required dependencies here
streamlit
fastapi
uvicorn
chromadb
google-generativeai
google-cloud-texttospeech
python-dotenv
